requirements.txt
discord.py
pymongo
requests
colorama
urllib3
python-dateutil
dnspython
python-dotenv